reorganised files in a directory structure
